feat: about 페이지 전용 메타데이터 구조 추가
차후 아날리시스 내 추적 및 SEO 최적화
